가게 리뷰 평점 scheduling 추가

+++ api/src/main/java/org/delivery/api/domain/store/controller/model/StoreResponse.java
@@ -22,7 +22,7 @@ public class StoreResponse {
 
     private StoreStatus status;
 
-    private double star;
+    private BigDecimal star;
 
     private String thumbnailUrl;
